Durability fix
Max durability seems to only be there for some items; for others, it seems to contain an unrelated value.

The items that do not have their durability on the item seem to always have a durability of 80. Where this is stored, or how the game determines when to use 80 vs the given value, is unknown.