mv rombot_updater.exe
Not sure why, but under Windows 11, trying to run it from the bin subfolder results in an error saying that it is in use by another program; this is false.

I can only assume this has something to do with security settings, but even run with elevated permissions this is still the case. It does not appear to be an issue with MicroMacro, either, as this same odd behavior can be seen directly from cmd.exe.

Simply moving the binary into the script's root resolves the issue, and will make it easier for users -- simply double-click the executable.